requirements.txt
# requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-api-python-client==2.110.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
cerebras_cloud_sdk>=0.1.0  # The correct package name from the documentation